requirements.txt
streamlit==1.25.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3
plotly==5.15.0
wordcloud==1.9.2
emoji==2.8.0
textblob==0.17.1
nltk==3.8.1
altair==5.0.1
networkx==3.1
google-generativeai